Update R730xd nic-configs and readme for p6p{1,2} to p4p{1,2} change
The 10G NICs for all the R730XDs in cloud15 were physically moved
so that they would be on the second NUMA node (the other NICs are
on the first NUMA node). As a result of this change the NICs are
presented as p4p1 instead of p6p1 (and similar).